pyproject.toml
[tool.poetry]
name = "Spatial-Representation-Analysis-of-Vision-Transformers-for-Satellite-Image-Classification"
version = "0.1.0"
description = "Analysis of Spatial Representations in Vision Transformers for EuroSAT"
authors = ["Thomas Fey-Grytnes <work@fey-grytnes.com>"]
readme = "README.md"
license = "MIT"
packages = [{include = "eurosat_vit_analysis"}]

[tool.poetry.dependencies]
python = "^3.11"
torch = "^2.2.0"
torchvision = "^0.17.0"
torchaudio = "^2.2.0"
timm = "^0.9.5"
dvc = {extras = ["s3"], version = "^3.0.0"} # Assuming S3 for remote storage
wandb = "^0.23.1"
hydra-core = "^1.3.2"
streamlit = "^1.26.0"
matplotlib = "^3.7.2"
seaborn = "^0.12.2"
scikit-learn = "^1.3.0"
pandas = "^2.0.3"
pillow = "^10.0.0"
pyyaml = "^6.0"
numpy = "<2.0"
peft = "^0.18.0"
transformers = "^4.57.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
pre-commit = "^3.4.0"
ruff = "^0.0.288"
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
# Enable Flake8, isort, and pyupgrade rules.
select = ["E", "F", "I", "UP"]
ignore = []

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.isort]
known-first-party = ["eurosat_vit_analysis"]
